added list on mobile rather than map (google chrome issue)
